handle the default datatype as no occurrence
The default datatype is not added to the listOfSimpleDatatypes, so in a example it was serialized as -1 (the result of indexOf), so I changed the datatypeID elements to have a minOccurs of 0 and a lack of occurrence means the default datatype. I didn't changed the logic for the globalComplexTypeGrammarID or globalAttributeDatatypeID. I am not exactly sure about them.